migrated from long to int key sizes

+++ RSA.h
@@ -4,7 +4,7 @@
 #define DEBUG 1
 
 #define SMS_SIZE            160
-#define PLAINTEXT_SIZE      (SMS_SIZE / sizeof(long))
+#define PLAINTEXT_SIZE      (SMS_SIZE / sizeof(int))
 #define CIPHERTEXT_SIZE     (SMS_SIZE)
 
 #include "Arduino.h"
@@ -15,8 +15,8 @@ class RSA {
     public:
         RSA();
         ~RSA();
-        void encrypt(char *plainText, char *chipherText, long *publicKey);
-        void decrypt(char *plainText, char *cipherText, long *privateKey);
+        void encrypt(char *plainText, char *chipherText, int *publicKey);
+        void decrypt(char *plainText, char *cipherText, int *privateKey);
         bool compare(char *arr1, char *arr2, int len);
 };
 extern RSA rsa;
